Use singed integers for sizes and indices
This undoes de653c9d873d5bc371d4b886e4a8ca8a97d00995, which was a
misguided attempt at consistency. It's preferable to use signed integers
since it makes math operations less problematic.

+++ include/Aulib/AudioDecoder.h
@@ -42,11 +42,11 @@ public:
     static AudioDecoder* decoderFor(SDL_RWops* rwops);
 
     bool isOpen() const;
-    size_t decode(float buf[], size_t len, bool& callAgain);
+    int decode(float buf[], int len, bool& callAgain);
 
     virtual bool open(SDL_RWops* rwops) = 0;
-    virtual unsigned getChannels() const = 0;
-    virtual unsigned getRate() const = 0;
+    virtual int getChannels() const = 0;
+    virtual int getRate() const = 0;
     virtual bool rewind() = 0;
     virtual float duration() const = 0;
     virtual bool seekToTime(float seconds) = 0;
@@ -54,7 +54,7 @@ public:
 protected:
     void setIsOpen(bool f);
 
-    virtual size_t doDecoding(float buf[], size_t len, bool& callAgain) = 0;
+    virtual int doDecoding(float buf[], int len, bool& callAgain) = 0;
 
 private:
     const std::unique_ptr<class AudioDecoder_priv> d;
