+++ include/Aulib/AudioResamplerSrc.h
@@ -32,9 +32,9 @@ public:
     ~AudioResamplerSRC() override;
 
 protected:
-    void doResampling(float dst[], const float src[], size_t& dstLen, size_t& srcLen) override;
+    void doResampling(float dst[], const float src[], int& dstLen, int& srcLen) override;
 
-    int adjustForOutputSpec(unsigned dstRate, unsigned srcRate, unsigned channels) override;
+    int adjustForOutputSpec(int dstRate, int srcRate, int channels) override;
 
 private:
     const std::unique_ptr<struct AudioResamplerSRC_priv> d;
